refactored blog editing code and also made docker not take up 70+ gb of cache

+++ app/admin/routes.py
@@ -17,45 +17,6 @@ from app.forms import *
 import app.util as util
 
 
-def sanitize_filename(filename):
-    filename = escape(secure_filename(filename))
-    filename = filename.replace("(", "").replace(")", "") # for Markdown parsing
-    return filename
-
-
-def validate_image(image):
-    header = image.read(512)
-    image.seek(0)
-    format = imghdr.what(None, header)
-    if not format:
-        return None
-    return "." + (format if format != "jpeg" else "jpg")
-
-
-def upload_images(images, path_before_filename) -> str:
-    for image in images:
-        if image.filename == "": # this happens when no image is submitted
-            continue
-
-        filename = sanitize_filename(image.filename)
-        if filename == "":
-            return "File name was reduced to atoms by sanitization."
-
-        file_ext = os.path.splitext(filename)[1]
-        invalid = \
-                file_ext not in current_app.config["IMAGE_UPLOAD_EXTENSIONS"] \
-                or (file_ext in current_app.config["IMAGE_UPLOAD_EXTENSIONS_CAN_VALIDATE"] \
-                and file_ext != validate_image(image.stream)) # imghdr can't check SVG; trustable since admin-only?
-        if invalid:
-            return "Invalid image. If it's another heic or webp im gonna lose my mind i swear to god i hate heic and webp theyre so annoying i hat"
-
-        path = os.path.join(path_before_filename, filename)
-        os.makedirs(path_before_filename, exist_ok=True)      # mkdir -p if not exist
-        image.save(path)                                      # replaces image if it's already there
-
-    return "success"
-
-
 @bp.route("/login", methods=["GET", "POST"])
 def login():
     form = LoginForm()
@@ -140,7 +101,7 @@ def create_blogpost():
             blogpage_id = int(request.args.get("blogpage_id", default=None))
         except Exception:
             return jsonify(redirect_url_abs=url_for("blog.index",
-                    flash=util.encode_URI_component("Good try."),
+                    flash=util.encode_URI_component("Nice try."),
                     _external=True))
 
         # automatically populate blogpage form field from query string if detected
@@ -157,40 +118,23 @@ def create_blogpost():
 
         post = Post(blogpage_id=request.form.get("blogpage_id"), title=request.form.get("title"),
                 subtitle=request.form.get("subtitle"), content=request.form.get("content"))
-        if post.subtitle == "": # standardize to None/NULL
-            post.subtitle = None
+        
         post.sanitize_title()
         db.session.add(post)
-        db.session.flush() # must do this before expand_image_markdown so post gets automatically assigned an id
-                           # must also do this to auto populate blogpage relationship from foreign key to use later
-                           # must also do add() before are_titles_unique(), as specified by are_titles_unique()
-
-        # check that title still exists after sanitization
-        if post.sanitized_title == "":
-            return jsonify(flash_message="Post must have alphanumeric characters in its title.")
-        # check that titles are unique
-        if not post.are_titles_unique():
-            return jsonify(flash_message="There is already a post with that title or sanitized title.")
-
-        # mark post as published and editable if creating on published blogpage
-        if not post.blogpage.unpublished:
-            post.published = True
-        else:
-            post.published = False
-
+        res = post.check_titles()
+        if res is not None:
+            return jsonify(flash_message=res)
+        post.add_timestamps(False, True)
         post.expand_image_markdown()
         db.session.commit()
 
         # upload images if any
-        try:
-            images_path = os.path.join(current_app.root_path,
+        images_path = os.path.join(current_app.root_path,
                     current_app.config["ROOT_TO_BLOGPAGE_STATIC"],
                     str(post.blogpage_id), "images", str(post.id))
-            res = upload_images(request.files.getlist("images"), images_path)
-            if not res == "success":
-                return jsonify(flash_message=res)
-        except Exception as e:
-            return jsonify(flash_message=f"File upload exception: {str(e)}")
+        res = upload_images(request.files.getlist("images"), images_path)
+        if res is not None:
+            return jsonify(flash_message=res)
 
         return jsonify(redirect_url_abs=url_for(f"blog.{post.blogpage_id}.post",
                 post_sanitized_title=post.sanitized_title, _external=True),
@@ -228,21 +172,21 @@ def edit_blogpost():
         post_id = int(request.args.get("post_id"))
     except Exception:
         return jsonify(redirect_url_abs=url_for("admin.search_blogpost", _external=True),
-                flash_message="Good try.")
+                flash_message="Nice try.")
 
     post = db.session.get(Post, post_id)
     if post is None:
         return jsonify(redirect_url_abs=url_for("admin.search_blogpost", _external=True),
                 flash_message="That post no longer exists. Did you hit the back button? Regret it, do you?")
-
-    images_path = os.path.join(current_app.root_path,
-            current_app.config["ROOT_TO_BLOGPAGE_STATIC"],
-            str(post.blogpage_id), "images", str(post.id))
-
+    
     form = EditBlogpostForm(obj=post) # pre-populate fields by name; again form must be created outside
     all_blogpages = db.session.query(Blogpage).order_by(Blogpage.ordering).all()
     form.blogpage_id.choices = [(blogpage.id, blogpage.title) for blogpage in all_blogpages if blogpage.writeable]
     form.content.data = post.collapse_image_markdown()
+    
+    images_path = os.path.join(current_app.root_path,
+            current_app.config["ROOT_TO_BLOGPAGE_STATIC"],
+            str(post.blogpage_id), "images", str(post.id))
     if os.path.exists(images_path) and os.path.isdir(images_path):
         images_choices = [(f, f) for f in os.listdir(images_path) \
                 if os.path.isfile(os.path.join(images_path, f))]
@@ -258,7 +202,7 @@ def edit_blogpost():
         if not form.validate():
             return jsonify(submission_errors=form.errors)
 
-        # handle post deletion (after confirmation button)
+        # handle post deletion
         if "delete" in request.form:
             db.session.delete(post)
             db.session.commit()
@@ -274,39 +218,23 @@ def edit_blogpost():
         old_blogpage_id = post.blogpage_id
         post.blogpage_id = request.form.get("blogpage_id")
         post.title = request.form.get("title")
-        post.sanitize_title()
-        db.session.flush()
-
-        if post.sanitized_title == "":
-            return jsonify(flash_message="Post must have alphanumeric characters in its title.")
-        if not post.are_titles_unique():
-            return jsonify(flash_message="There is already a post with that title or sanitized title.")
-
         post.subtitle = request.form.get("subtitle")
-        if post.subtitle == "":
-            post.subtitle = None
         post.content = request.form.get("content")
+        
+        post.sanitize_title()
+        res = post.check_titles()
+        if res is not None:
+            return jsonify(flash_message=res)
+        post.add_timestamps(request.form.get("remove_edited_timestamps"), # todo make sure works
+                request.form.get("dont_update_edited_timestamp"))
+        post.expand_image_markdown()
+        db.session.commit()
 
-        # update edited time if editing a published post
-        if post.published:
-            if request.form.get("remove_edited_timestamps"):
-                post.edited_timestamp = None
-            elif request.form.get("dont_update_edited_timestamp"):
-                pass
-            else:
-                post.edited_timestamp = datetime.now(timezone.utc)
-            # unpublish if moving back to backrooms
-            if post.blogpage.unpublished:
-                post.published = False
-                post.edited_timestamp = None
-        else:
-            # mark post as published and editable if not published and moving to published blogpage
-            if not post.blogpage.unpublished:
-                post.published = True
-            # keep updating created time instead of updated time if not published
-            post.timestamp = datetime.now(timezone.utc)
-        post.expand_image_markdown() # must be done after post.blogpage_id is finalized for image markdown updating!
-
+        # upload images if any
+        res = upload_images(request.files.getlist("images"), images_path)
+        if res is not None:
+            return jsonify(flash_message=res)
+        
         # delete images if any
         try:
             for image in request.form.getlist("delete_images"):
@@ -316,42 +244,31 @@ def edit_blogpost():
             if os.path.exists(images_path) and len(os.listdir(images_path)) == 0 and os.path.isdir(images_path):
                 shutil.rmtree(images_path)
         except Exception as e:
-            return jsonify(flash_message=f"File delete exception: {str(e)}")
+            return jsonify(flash_message=f"Image delete exception: {str(e)}")
 
-        # delete unused images if that is checked; do this after deleting normally
-        # but before uploading to make sure newly-uploaded images aren't immediately scrapped
-        if request.form.get("delete_unused_images"):
+        # delete unused images if applicable; do this after deleting normally but before uploading
+        # to make sure newly-uploaded images aren't immediately scrapped
+        if request.form.get("delete_unused_images") and os.path.exists(images_path):
             try:
-                if os.path.exists(images_path):
-                    images = os.listdir(images_path)
-                    # todo check without extension?
-                    for image in images:
-                        file_ext = os.path.splitext(image)[1]
-                        if file_ext in current_app.config["IMAGE_UPLOAD_EXTENSIONS_CAN_DELETE_UNUSED"] \
-                                and image not in post.content:
-                            os.remove(os.path.join(images_path, image))
+                images = os.listdir(images_path)
+                for image in images:
+                    file_ext = os.path.splitext(image)[1]
+                    if file_ext in current_app.config["IMAGE_UPLOAD_EXTENSIONS_CAN_DELETE_UNUSED"] \
+                            and image not in post.content:
+                        os.remove(os.path.join(images_path, image))
 
-                    if len(os.listdir(images_path)) == 0 and os.path.isdir(images_path):
-                        shutil.rmtree(images_path)
+                if len(os.listdir(images_path)) == 0 and os.path.isdir(images_path):
+                    shutil.rmtree(images_path)
             except Exception as e:
-                return jsonify(flash_message=f"File delete unused exception: {str(e)}")
+                return jsonify(flash_message=f"Image delete unused exception: {str(e)}")
         
-        # upload images if any
-        try:
-            res = upload_images(request.files.getlist("images"), images_path)
-            if not res == "success":
-                return jsonify(flash_message=res)
-        except Exception as e:
-            return jsonify(flash_message=f"File upload exception: {str(e)}")
-
         # move images if moving blogpost
         if post.blogpage_id != old_blogpage_id:
             if os.path.exists(images_path):
                 try:
                     shutil.move(images_path, os.path.join(os.path.split(images_path)[0], str(post.id)))
                 except Exception as e:
-                    return jsonify(flash_message=f"File move exception: {str(e)}")
-        db.session.commit()
+                    return jsonify(flash_message=f"Image move exception: {str(e)}")
 
         return jsonify(redirect_url_abs=url_for(f"blog.{post.blogpage_id}.post",
                 post_sanitized_title=post.sanitized_title, _external=True),
@@ -378,7 +295,6 @@ def change_admin_password():
         # check given password
         if user is None or not user.check_password(request.form.get("old_password")):
             return jsonify(flash_message="Old password is not correct.")
-
         # check new passwords are identical
         if request.form.get("new_password_1") != request.form.get("new_password_2"):
             return jsonify(flash_message="New passwords do not match.")
@@ -404,3 +320,50 @@ def logout():
                     flash_message="Mischief managed.")
 
     return jsonify(flash_message="Mischief managed.")
+
+
+###################################################################################################
+#                                         Helper functions                                        # 
+###################################################################################################
+
+
+def sanitize_filename(filename):
+    filename = escape(secure_filename(filename))
+    filename = filename.replace("(", "").replace(")", "") # for Markdown parsing
+    return filename
+
+
+def validate_image(image):
+    header = image.read(512)
+    image.seek(0)
+    format = imghdr.what(None, header)
+    if not format:
+        return None
+    return "." + (format if format != "jpeg" else "jpg")
+
+
+def upload_images(images, images_path) -> str:
+    try:
+        for image in images:
+            if image.filename == "": # this happens when no image is submitted
+                continue
+
+            filename = sanitize_filename(image.filename)
+            if filename == "":
+                return "Image name was reduced to atoms by sanitization."
+
+            file_ext = os.path.splitext(filename)[1]
+            invalid = \
+                    file_ext not in current_app.config["IMAGE_UPLOAD_EXTENSIONS"] \
+                    or (file_ext in current_app.config["IMAGE_UPLOAD_EXTENSIONS_CAN_VALIDATE"] \
+                    and file_ext != validate_image(image.stream)) # imghdr can't check SVG; trustable since admin-only?
+            if invalid:
+                return "Invalid image. If it's another heic or webp im gonna lose my mind i swear to god i hate heic and webp theyre so annoying i hat"
+
+            path = os.path.join(images_path, filename)
+            os.makedirs(images_path, exist_ok=True)               # mkdir -p if not exist
+            image.save(path)                                      # replaces image if it's already there
+    except Exception as e:
+        return f"Image upload exception: {str(e)}"
+    
+    return None
